+++ test/support/classifier.py
@@ -13,6 +13,7 @@ X_test = X[100:]
 y_test = y[100:]
 
 model = xgb.XGBClassifier()
-model.fit(X_train, y_train)
+# model.fit(X_train, y_train)
+model.fit(X_train, y_train, eval_set=[(X_test, y_test)], early_stopping_rounds=5, verbose=True)
 # print(model.predict_proba(X_test))
 print(model.feature_importances_)