+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Friendlier message when XGBoost not found
 - Free memory when objects are destroyed
 - Added `Ranker`
+- Added early stopping to Scikit-Learn API
 
 ## 0.1.1